// bedrock-types.ts
export type ItemLegacy = {
  network_id: number;
  count?: number;
  metadata?: number;
  block_runtime_id?: number;
  extra?:
    | ItemExtraDataWithBlockingTick
    | ItemExtraDataWithoutBlockingTick
    | undefined;
};
/**
 * An "ItemStack" here represents an Item instance. You can think about it like a pointer
 * to an item class. The data for the class gets updated with the data in the `item` field
 * As of 1.16.220, now functionally the same as `Item` just without an extra boolean when
 * server auth inventories is disabled.
 */
export type Item = {
  network_id: number;
  count?: number;
  metadata?: number;
  has_stack_id?: number;
  stack_id?: number | undefined;
  block_runtime_id?: number;
  extra?:
    | ItemExtraDataWithBlockingTick
    | ItemExtraDataWithoutBlockingTick
    | undefined;
};
export type vec3i = {
  x: number;
  y: number;
  z: number;
};
export type vec3li = {
  x: number;
  y: number;
  z: number;
};
export type vec3u = {
  x: number;
  y: number;
  z: number;
};
export type vec3f = {
  x: number;
  y: number;
  z: number;
};
export type vec2f = {
  x: number;
  z: number;
};
// ...
export type BlockCoordinates = {
  x: number;
  y: number;
  z: number;
};
// ...
export type TransactionUseItem = {
  /** ActionType is the type of the UseItem inventory transaction. It is one of the action types found above, and specifies the way the player interacted with the block. */
  action_type: "click_block" | "click_air" | "break_block" | "attack";
  /** TriggerType is the type of the trigger that caused the inventory transaction. It is one of the trigger types found in the constants above. If TriggerType is TriggerTypePlayerInput, the transaction is from the initial input of the player. If it is TriggerTypeSimulationTick, the transaction is from a simulation tick when the player is holding down the input. */
  trigger_type: "unknown" | "player_input" | "simulation_tick";
  /** BlockPosition is the position of the block that was interacted with. This is only really a correct block position if ActionType is not UseItemActionClickAir. */
  block_position: BlockCoordinates;
  /** BlockFace is the face of the block that was interacted with. When clicking the block, it is the face clicked. When breaking the block, it is the face that was last being hit until the block broke. */
  face: number;
  /** HotBarSlot is the hot bar slot that the player was holding while clicking the block. It should be used to ensure that the hot bar slot and held item are correctly synchronised with the server. */
  hotbar_slot: number;
  /** HeldItem is the item that was held to interact with the block. The server should check if this item is actually present in the HotBarSlot. */
  held_item: Item;
  /** Position is the position of the player at the time of interaction. For clicking a block, this is the position at that time, whereas for breaking the block it is the position at the time of breaking. */
  player_pos: vec3f;
  /** ClickedPosition is the position that was clicked relative to the block's base coordinate. It can be used to find out exactly where a player clicked the block. */
  click_pos: vec3f;
  /** BlockRuntimeID is the runtime ID of the block that was clicked. It may be used by the server to verify that the player's world client-side is synchronised with the server's. */
  block_runtime_id: number;
  /** ClientPrediction is the client's prediction on the output of the transaction. */
  client_prediction: "failure" | "success";
};
/**
 * Actions is a list of actions that took place, that form the inventory transaction together. Each of
 * these actions hold one slot in which one item was changed to another. In general, the combination of
 * all of these actions results in a balanced inventory transaction. This should be checked to ensure that
 * no items are cheated into the inventory.
 */
export type TransactionActions = {
  source_type:
    | "container"
    | "global"
    | "world_interaction"
    | "creative"
    | "craft_slot"
    | "craft";
  inventory_id?: WindowIDVarint;
  action?: number;
  flags?: number;
  slot: number;
  old_item: Item;
  new_item: Item;
}[];
/**
 * The Minecraft bedrock inventory system was refactored, but not all inventory actions use the new packet.
 * This data structure holds actions that have not been updated to the new system.
 */
export type TransactionLegacy = {
  /** LegacyRequestID is an ID that is only non-zero at times when sent by the client. The server should always send 0 for this. When this field is not 0, the LegacySetItemSlots slice below will have values in it. LegacyRequestID ties in with the ItemStackResponse packet. If this field is non-0, the server should respond with an ItemStackResponse packet. Some inventory actions such as dropping an item out of the hotbar are still one using this packet, and the ItemStackResponse packet needs to tie in with it. */
  legacy_request_id: number;
  /** `legacy_transactions` are only present if the LegacyRequestID is non-zero. These item slots inform the server of the slots that were changed during the inventory transaction, and the server should send back an ItemStackResponse packet with these slots present in it. (Or false with no slots, if rejected.) */
  legacy_transactions?: {
    container_id: number;
    changed_slots: {
      slot_id: number;
    }[];
  }[];
};
// ...
export type Action =
  | "start_break"
  | "abort_break"
  | "stop_break"
  | "get_updated_block"
  | "drop_item"
  | "start_sleeping"
  | "stop_sleeping"
  | "respawn"
  | "jump"
  | "start_sprint"
  | "stop_sprint"
  | "start_sneak"
  | "stop_sneak"
  | "creative_player_destroy_block"
  | "dimension_change_ack"
  | "start_glide"
  | "stop_glide"
  | "build_denied"
  | "crack_break"
  | "change_skin"
  | "set_enchatnment_seed"
  | "swimming"
  | "stop_swimming"
  | "start_spin_attack"
  | "stop_spin_attack"
  | "interact_block"
  | "predict_break"
  | "continue_break"
  | "start_item_use_on"
  | "stop_item_use_on"
  | "handled_teleport"
  | "missed_swing"
  | "start_crawling"
  | "stop_crawling"
  | "start_flying"
  | "stop_flying"
  | "received_server_data"
  | "start_using_item";
/**
 * Source and Destination point to the source slot from which Count of the item stack were taken and the
 * destination slot to which this item was moved.
 */
export type StackRequestSlotInfo = {
  /** ContainerID is the ID of the container that the slot was in. */
  slot_type: FullContainerName;
  /** Slot is the index of the slot within the container with the ContainerID above. */
  slot: number;
  /** StackNetworkID is the unique stack ID that the client assumes to be present in this slot. The server must check if these IDs match. If they do not match, servers should reject the stack request that the action holding this info was in. */
  stack_id: number;
};
/**
 * ItemStackRequest is sent by the client to change item stacks in an inventory. It is essentially a
 * replacement of the InventoryTransaction packet added in 1.16 for inventory specific actions, such as moving
 * items around or crafting. The InventoryTransaction packet is still used for actions such as placing blocks
 * and interacting with entities.
 */
export type ItemStackRequest = {
  /** RequestID is a unique ID for the request. This ID is used by the server to send a response for this specific request in the ItemStackResponse packet. */
  request_id: number;
  actions: {
    type_id:
      | "take"
      | "place"
      | "swap"
      | "drop"
      | "destroy"
      | "consume"
      | "create"
      | "place_in_container"
      | "take_out_container"
      | "lab_table_combine"
      | "beacon_payment"
      | "mine_block"
      | "craft_recipe"
      | "craft_recipe_auto"
      | "craft_creative"
      | "optional"
      | "craft_grindstone_request"
      | "craft_loom_request"
      | "non_implemented"
      | "results_deprecated";
    count?: number;
    source?: StackRequestSlotInfo;
    destination?: StackRequestSlotInfo;
    randomly?: boolean;
    result_slot_id?: number;
    primary_effect?: number;
    secondary_effect?: number;
    hotbar_slot?: number;
    predicted_durability?: number;
    network_id?: number;
    recipe_network_id?: number;
    times_crafted?: number;
    times_crafted_2?: number;
    ingredients?: RecipeIngredient[];
    item_id?: number;
    filtered_string_index?: number;
    cost?: number;
    pattern?: string;
    result_items?: ItemLegacy[];
  }[];
  custom_names: string[];
  cause:
    | "chat_public"
    | "chat_whisper"
    | "sign_text"
    | "anvil_text"
    | "book_and_quill_text"
    | "command_block_text"
    | "block_actor_data_text"
    | "join_event_text"
    | "leave_event_text"
    | "slash_command_chat"
    | "cartography_text"
    | "kick_command"
    | "title_command"
    | "summon_command";
};
// ...
export type WindowIDVarint =
  | "inventory"
  | "first"
  | "last"
  | "offhand"
  | "armor"
  | "creative"
  | "hotbar"
  | "fixed_inventory"
  | "ui"
  | "drop_contents"
  | "beacon"
  | "trading_output"
  | "trading_use_inputs"
  | "trading_input_2"
  | "trading_input_1"
  | "enchant_output"
  | "enchant_material"
  | "enchant_input"
  | "anvil_output"
  | "anvil_result"
  | "anvil_material"
  | "container_input"
  | "crafting_use_ingredient"
  | "crafting_result"
  | "crafting_remove_ingredient"
  | "crafting_add_ingredient"
  | "none";
// ...
export type packet_player_auth_input = {
  /** Pitch that the player reports it has. */
  pitch: number;
  /** Yaw that player reports it has. */
  yaw: number;
  /** Position holds the position that the player reports it has. */
  position: vec3f;
  /** MoveVector is a Vec2 that specifies the direction in which the player moved, as a combination of X/Z values which are created using the WASD/controller stick state. */
  move_vector: vec2f;
  /** HeadYaw is the horizontal rotation of the head that the player reports it has. */
  head_yaw: number;
  /** InputData is a combination of bit flags that together specify the way the player moved last tick. It is a combination of the flags above. */
  input_data: InputFlag;
  /** InputMode specifies the way that the client inputs data to the screen. It is one of the constants that may be found above. */
  input_mode: "unknown" | "mouse" | "touch" | "game_pad" | "motion_controller";
  /** PlayMode specifies the way that the player is playing. The values it holds, which are rather random, may be found above. */
  play_mode:
    | "normal"
    | "teaser"
    | "screen"
    | "viewer"
    | "reality"
    | "placement"
    | "living_room"
    | "exit_level"
    | "exit_level_living_room"
    | "num_modes";
  /** InteractionModel is a constant representing the interaction model the player is using. */
  interaction_model: "touch" | "crosshair" | "classic";
  /** interact_rotation is the rotation the player is looking that they intend to use for interactions. This is only different to Pitch and Yaw in cases such as VR or when custom cameras being used. */
  interact_rotation: vec2f;
  /** Tick is the server tick at which the packet was sent. It is used in relation to CorrectPlayerMovePrediction. */
  tick: bigint;
  /** Delta was the delta between the old and the new position. There isn't any practical use for this field as it can be calculated by the server itself. */
  delta: vec3f;
  transaction?: {
    legacy: TransactionLegacy;
    actions: TransactionActions;
    data: TransactionUseItem;
  };
  item_stack_request?: ItemStackRequest;
  vehicle_rotation?: vec2f;
  predicted_vehicle?: bigint;
  block_action?: {
    action: Action;
    position?: vec3i;
    face?: number;
  }[];
  analogue_move_vector: vec2f;
  camera_orientation: vec3f;
  raw_move_vector: vec2f;
};
export type InputFlag = {
  ascend?: boolean;
  descend?: boolean;
  north_jump?: boolean;
  jump_down?: boolean;
  sprint_down?: boolean;
  change_height?: boolean;
  jumping?: boolean;
  auto_jumping_in_water?: boolean;
  sneaking?: boolean;
  sneak_down?: boolean;
  up?: boolean;
  down?: boolean;
  left?: boolean;
  right?: boolean;
  up_left?: boolean;
  up_right?: boolean;
  want_up?: boolean;
  want_down?: boolean;
  want_down_slow?: boolean;
  want_up_slow?: boolean;
  sprinting?: boolean;
  ascend_block?: boolean;
  descend_block?: boolean;
  sneak_toggle_down?: boolean;
  persist_sneak?: boolean;
  start_sprinting?: boolean;
  stop_sprinting?: boolean;
  start_sneaking?: boolean;
  stop_sneaking?: boolean;
  start_swimming?: boolean;
  stop_swimming?: boolean;
  start_jumping?: boolean;
  start_gliding?: boolean;
  stop_gliding?: boolean;
  item_interact?: boolean;
  block_action?: boolean;
  item_stack_request?: boolean;
  handled_teleport?: boolean;
  emoting?: boolean;
  missed_swing?: boolean;
  start_crawling?: boolean;
  stop_crawling?: boolean;
  start_flying?: boolean;
  stop_flying?: boolean;
  received_server_data?: boolean;
  client_predicted_vehicle?: boolean;
  paddling_left?: boolean;
  paddling_right?: boolean;
  block_breaking_delay_enabled?: boolean;
  horizontal_collision?: boolean;
  vertical_collision?: boolean;
  down_left?: boolean;
  down_right?: boolean;
  start_using_item?: boolean;
  camera_relative_movement_enabled?: boolean;
  rot_controlled_by_move_direction?: boolean;
  start_spin_attack?: boolean;
  stop_spin_attack?: boolean;
  hotbar_only_touch?: boolean;
  jump_released_raw?: boolean;
  jump_pressed_raw?: boolean;
  jump_current_raw?: boolean;
  sneak_released_raw?: boolean;
  sneak_pressed_raw?: boolean;
  sneak_current_raw?: boolean;
};